#27 - moves balance card component and its test to feature module
