+++ app/src/main/java/br/com/rbrthmn/di/AppModule.kt
@@ -20,10 +20,10 @@
 
 package br.com.rbrthmn.di
 
+import br.com.rbrthmn.home.ui.components.balancecard.BalanceCardContract
+import br.com.rbrthmn.home.ui.components.balancecard.BalanceCardViewModel
 import br.com.rbrthmn.ui.financialcompanion.screens.home.HomeScreenContract
 import br.com.rbrthmn.ui.financialcompanion.screens.home.HomeScreenViewModel
-import br.com.rbrthmn.ui.financialcompanion.screens.home.components.balancecard.BalanceCardContract
-import br.com.rbrthmn.ui.financialcompanion.screens.home.components.balancecard.BalanceCardViewModel
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.creditcardbillscard.CreditCardBillsCardContract
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.creditcardbillscard.CreditCardBillsCardViewModel
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.lastmonthdifferencecard.LastMonthDifferenceCardContract
