+++ core/ui/src/main/res/values/strings.xml
@@ -10,6 +10,10 @@
     <string name="bank_hint">Banco</string>
     <string name="drop_down_arrow_icon_description">Ícone de flecha para baixo</string>
 
+    <string name="cancel_button">Cancelar</string>
+    <string name="save_button">Salvar</string>
+    <string name="colon" translatable="false">:</string>
+
     <string name="current_month_button">Mês atual</string>
     <string name="apply_date_button">Aplicar</string>
